Remove deprecated `EdgeAccount.pluginData`
