+++ src/core/account/account-api.js
@@ -11,7 +11,6 @@ import {
   type EdgeCurrencyWallet,
   type EdgeDataStore,
   type EdgeLobby,
-  type EdgePluginData,
   type EdgePluginMap,
   type EdgeRateCache,
   type EdgeSwapConfig,
@@ -48,7 +47,7 @@ import { type ApiInput } from '../root-pixie.js'
 import { makeStorageWalletApi } from '../storage/storage-api.js'
 import { fetchSwapQuote } from '../swap/swap-api.js'
 import { changeWalletStates } from './account-files.js'
-import { makeDataStoreApi, makePluginDataApi } from './data-store-api.js'
+import { makeDataStoreApi } from './data-store-api.js'
 import { makeLobbyApi } from './lobby-api.js'
 import { CurrencyConfig, SwapConfig } from './plugin-api.js'
 
@@ -75,7 +74,6 @@ export function makeAccountApi(ai: ApiInput, accountId: string): EdgeAccount {
   // Specialty API's:
   const rateCache = makeExchangeCache(ai)
   const dataStore = makeDataStoreApi(ai, accountId)
-  const pluginData = makePluginDataApi(dataStore)
   const storageWalletApi = makeStorageWalletApi(ai, accountWalletInfo)
 
   function lockdown() {
@@ -355,9 +353,6 @@ export function makeAccountApi(ai: ApiInput, accountId: string): EdgeAccount {
     },
     get exchangeCache(): EdgeRateCache {
       return rateCache
-    },
-    get pluginData(): EdgePluginData {
-      return pluginData
     },
     async getExchangeQuote(request: EdgeSwapRequest): Promise<EdgeSwapQuote> {
       deprecate('EdgeAccount.getExchangeQuote', 'EdgeAccount.fetchSwapQuote')
